Removes the idiotic decision of myself to store ports in signed shorts.

+++ app/src/main/java/de/kuschku/quasseldroid_ng/ui/chat/ChatActivity.kt
@@ -106,7 +106,7 @@ class ChatActivity : ServiceBoundActivity() {
             finish()
           } else {
             backendValue.connectUnlessConnected(
-              SocketAddress(account.host, account.port.toShort()),
+              SocketAddress(account.host, account.port),
               account.user,
               account.pass,
               true
